SorrTask606 Follow up Get email from dropcontact (#616)
* get email from dropcontact

* Update
